[#156] Fix: 거절 누르면 초대받은 대시보드만 초기화 되게 바꿈
