+++ components/Table/InviteDashBoard.tsx
@@ -1,116 +1,33 @@
-import API from '@/apis/api';
 import useDeviceType from '@/hooks/useDeviceType';
+import useInviteDashBoard from '@/hooks/useInviteDashBoard';
 import search from '@/public/assets/icons/Search.svg';
 import NoContent from '@/public/assets/images/NoContent.png';
 import { fontStyle } from '@/styles/fontStyle';
 import { onMobile, onTablet } from '@/styles/mediaQuery';
 import { COLORS } from '@/styles/palettes';
 import Image from 'next/image';
-import { ChangeEvent, KeyboardEvent, useEffect, useRef, useState } from 'react';
 import InfiniteScroll from 'react-infinite-scroller';
 import styled, { css } from 'styled-components';
 import TwinButton from '../common/Button/TwinButton';
 
-interface GetInvitationListProps {
-  id: number;
-  inviter: {
-    id: number;
-    email: string;
-    nickname: string;
-  };
-  teamId: string;
-  dashboard: {
-    title: string;
-    id: number;
-  };
-  invitee: {
-    nickname: string;
-    email: string;
-    id: number;
-  };
-  inviteAccepted: boolean;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface GetAcceptProps {
-  acceptid: number;
-  accept: boolean;
-}
-
-interface InviteDashBoardProps {
+export interface InviteDashBoardProps {
   refresh: () => void;
   refreshToFirst: () => void;
 }
 
 function InviteDashBoard({ refresh, refreshToFirst }: InviteDashBoardProps) {
-  const [dataSource, setDataSource] = useState<GetInvitationListProps[]>([]);
-  const [hasMore, setHasMore] = useState(true);
-  const [searchText, setSearchText] = useState('');
-  const [cursor, setCursor] = useState(0);
   const windowSize = useDeviceType();
   const width = windowSize === 'pc' || windowSize === 'tablet';
-  const InviteContainerRef = useRef<HTMLDivElement>(null);
-
-  const fetchHasMore = () => {
-    if (cursor) {
-      if (dataSource.length !== 0) {
-        handleLoadMore();
-      }
-    } else {
-      setHasMore((prev) => !prev);
-    }
-  };
-
-  const handleLoadMore = async () => {
-    const item = await API.invitations.getInvitationList({ cursorId: cursor });
-    setDataSource((prev) => [...prev, ...item.invitations]);
-    setCursor(item.cursorId as number);
-  };
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setSearchText(e.target.value);
-  };
-
-  const getItems = async () => {
-    const item = await API.invitations.getInvitationList({});
-    setDataSource(item.invitations);
-    setCursor(item.cursorId as number);
-  };
-
-  const handleAccept = async ({ acceptid, accept }: GetAcceptProps) => {
-    await API.invitations.responseInvitation({ invitationId: acceptid, inviteAccepted: accept });
-    setDataSource(dataSource.filter((item) => item.id !== acceptid));
-    refresh();
-    refreshToFirst();
-  };
-
-  const handleSearch = async (searchText: string) => {
-    const searchItem = await API.invitations.getInvitationList({ title: searchText });
-    if (searchItem.invitations.length > 0) {
-      setDataSource(searchItem.invitations);
-      setCursor(searchItem.cursorId as number);
-    } else {
-      alert('검색된 데이터가 없습니다.');
-      getItems();
-      setSearchText('');
-    }
-  };
-
-  const handleKeyDown = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      e.preventDefault();
-      handleSearch(searchText);
-    }
-  };
-
-  useEffect(() => {
-    getItems();
-    if (InviteContainerRef.current) {
-      InviteContainerRef.current.scrollTop = 0;
-      setHasMore((prev) => !prev);
-    }
-  }, [refresh]);
+  const {
+    dataSource,
+    searchText,
+    handleChange,
+    handleKeyDown,
+    InviteContainerRef,
+    fetchHasMore,
+    hasMore,
+    handleAccept,
+  } = useInviteDashBoard({ refresh, refreshToFirst });
 
   return (
     <StyledDiv $data={dataSource}>
